auto-map/requirements-zmq.txt
PyQt5>=5.15.0
pyzmq>=22.0.0
numpy>=1.21.0
Pillow>=8.0.0
bluesky-queueserver-api>=0.0.8
pyepics>=3.5.0 